requirements.txt
flask
python-barcode
pillow
opencv-python
pyzbar
psycopg2-binary
reportlab
python-dotenv